refactor: extract small app-module pieces to feature/core modules
Moves reminder Koin registrations to logic-reminder, contact/file
pickers to platform-common and feature-birthday, keeping app module
free of logic that belongs closer to its owning feature.
