+++ logic/logic-reminder/src/main/kotlin/com/github/naz013/logic/reminder/KoinModule.kt
@@ -5,17 +5,22 @@ import com.github.naz013.logic.reminder.occurrence.CalculateReminderOccurrencesU
 import com.github.naz013.logic.reminder.occurrence.ReminderOccurrenceCalculatorFactoryV2
 import com.github.naz013.logic.reminder.query.GetActiveRemindersV2ByGroupIdUseCase
 import com.github.naz013.logic.reminder.query.ResolveReminderV2NotificationSettingsUseCase
+import com.github.naz013.logic.reminder.scheduling.EventDateTimeCalculatorV2
 import com.github.naz013.logic.reminder.usecase.ActivateReminderUseCase
+import com.github.naz013.logic.reminder.usecase.CompleteReminderUseCase
 import com.github.naz013.logic.reminder.usecase.CompleteRelatedGoogleTaskUseCase
 import com.github.naz013.logic.reminder.usecase.DeactivateReminderUseCase
 import com.github.naz013.logic.reminder.usecase.DeleteAllReminderUseCase
 import com.github.naz013.logic.reminder.usecase.DeleteReminderUseCase
 import com.github.naz013.logic.reminder.usecase.MoveReminderToArchiveUseCase
 import com.github.naz013.logic.reminder.usecase.PauseReminderUseCase
+import com.github.naz013.logic.reminder.usecase.ResumeReminderUseCase
 import com.github.naz013.logic.reminder.usecase.SaveOneTimeReminderUseCaseImpl
 import com.github.naz013.logic.reminder.usecase.SaveReminderUseCase
+import com.github.naz013.logic.reminder.usecase.SkipReminderUseCase
 import com.github.naz013.logic.reminder.usecase.StartLocationTrackingUseCase
 import com.github.naz013.logic.reminder.usecase.StopLocationTrackingUseCase
+import com.github.naz013.logic.reminder.usecase.ToggleReminderStateUseCase
 import com.github.naz013.logic.reminder.usecase.UpdatePermanentReminderNotificationUseCase
 import com.github.naz013.logic.reminder.usecase.google.SaveReminderToGoogleCalendarUseCase
 import com.github.naz013.logic.reminder.usecase.google.SaveReminderToGoogleTasksUseCase
@@ -51,4 +56,10 @@ val logicReminderModule = module {
   factoryOf(::SaveReminderToGoogleCalendarUseCase)
 
   factory { SaveOneTimeReminderUseCaseImpl(get(), get(), get()) as SaveOneTimeReminderUseCase }
+
+  factoryOf(::ResumeReminderUseCase)
+  factoryOf(::CompleteReminderUseCase)
+  factoryOf(::SkipReminderUseCase)
+  factoryOf(::ToggleReminderStateUseCase)
+  factory { EventDateTimeCalculatorV2(get(), get()) }
 }
